+++ sonic-xcvrd/xcvrd/dom/dom_mgr.py
@@ -16,10 +16,10 @@ try:
     from natsort import natsorted
     from swsscommon import swsscommon
 
-    from . import xcvrd
-    from .xcvrd_utilities import sfp_status_helper
-    from .xcvrd_utilities.xcvr_table_helper import *
-    from .xcvrd_utilities import port_event_helper
+    from xcvrd import xcvrd
+    from xcvrd.xcvrd_utilities import sfp_status_helper
+    from xcvrd.xcvrd_utilities.xcvr_table_helper import *
+    from xcvrd.xcvrd_utilities import port_event_helper
 except ImportError as e:
     raise ImportError(str(e) + " - required module not found in dom_mgr.py")
